Optional virtualbox install and version (#2)
This patch allows you to specify an optional version (i.e. name) for the virtualbox install on Debian systems.  This is necessary if you are using the oracle repos and are installing e.g. virtualbox-5.1.  You can have both virtualbox-5.1 and virtualbox (from the default repos) installed, but they will conflict.

This also allows you to override the install by choosing false (e.g. if you are installing it from another role).